+++ tests/testthat/test-loadData.R
@@ -3,6 +3,7 @@ context("loadData")
 test_that("baseUrl Input validation works", {
   expect_error(downloadParseSynopticData(baseUrl = "https://www.purpleair.com/DUMMY"))
   expect_error(pas_load(baseUrl = "https://www.purpleair.com/DUMMY"))
+  expect_error(pas_loadLatest(baseUrl = "https://www.purpleair.com/DUMMY"))
 })
 
 test_that("pwfsl loads correct class", {